requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
certifi==2025.11.12
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.127.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
orjson==3.11.5
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.5
python-jose==3.5.0
python-multipart==0.0.21
rsa==4.9.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
